Move pipeline to appropriate file
